Changed order quantity back to integer

+++ molveno/restaurant/models.py
@@ -24,7 +24,7 @@ class Inventory(models.Model):
     unit_price = models.DecimalField(max_digits=5, decimal_places=2)
     minimum_quantity = models.DecimalField(max_digits=5, decimal_places=2, default=0)
     current_stock = models.DecimalField(max_digits=5, decimal_places=2, default=0)
-    order_quantity = models.DecimalField(max_digits=5, decimal_places=2, default=0)
+    order_quantity = models.IntegerField(default=0)
 
     def __str__(self):
         return self.description
